% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra1}
\alias{dijkstra1}
\title{Dijkstra's algorithm}
\usage{
dijkstra1(graph, init_node)
}
\arguments{
\item{graph}{a data frame}

\item{init_node}{a numeric scalar}
}
\value{
a numeric vector, with the shortest distance to each node
}
\description{
find the shortest path algorithm in a graph
}
\details{
shortest path algorithm in a graph
}
