% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{a numeric scalar or integer}

\item{b}{a numeric scalar or integer}
}
\value{
a integer number
}
\description{
find the greatest common divisor of two integers
}
\details{
find the greatest common divisor of two integers
}
\references{
\url{https:\\en.wikipedia.org/wiki/Euclidean_algorithm.}
}
