\name{proyect3-package}
\docType{package}
\title{
\packageTitle{proyect3}
}
\description{
\packageDescription{proyect3}
}
\details{

The DESCRIPTION file
\packageDESCRIPTION{proyect3}
\packageIndices{proyect3}
Package contains of implementations of two algorithm. Dijkstra is for finding the shortest path from the initial node and Euclidean is for calculating greatest common divisior of two numbers. 
}
\author{ packageAuthor {Jorge Villar Fuentes}

Maintainer packageMaintainer{Jorge Villar Fuentes}
}

\keyword{ GCD, Euclidean Algorithm }
\seealso{https://en.wikipedia.org/wiki/Greatest_common_divisor.
https://www.youtube.com/watch?v=gdmfOwyQlcI.}
\examples{euclidean(100, 1000)
dijkstra(wiki_graph, 1)
}